Add integration tests for points
