+++ tests/test_integration_client.py
@@ -120,82 +120,6 @@ class TestClient(TestCase):
         schedule = client.season_schedule(season_end_year=2019)
         self.assertIsNotNone(schedule)
 
-    def test_2001_player_season_totals(self):
-        player_season_totals = client.players_season_totals(season_end_year=2001)
-        self.assertIsNotNone(player_season_totals)
-
-    def test_2002_player_season_totals(self):
-        player_season_totals = client.players_season_totals(season_end_year=2002)
-        self.assertIsNotNone(player_season_totals)
-
-    def test_2003_player_season_totals(self):
-        player_season_totals = client.players_season_totals(season_end_year=2003)
-        self.assertIsNotNone(player_season_totals)
-
-    def test_2004_player_season_totals(self):
-        player_season_totals = client.players_season_totals(season_end_year=2004)
-        self.assertIsNotNone(player_season_totals)
-
-    def test_2005_player_season_totals(self):
-        player_season_totals = client.players_season_totals(season_end_year=2005)
-        self.assertIsNotNone(player_season_totals)
-
-    def test_2006_player_season_totals(self):
-        player_season_totals = client.players_season_totals(season_end_year=2006)
-        self.assertIsNotNone(player_season_totals)
-
-    def test_2007_player_season_totals(self):
-        player_season_totals = client.players_season_totals(season_end_year=2007)
-        self.assertIsNotNone(player_season_totals)
-
-    def test_2008_player_season_totals(self):
-        player_season_totals = client.players_season_totals(season_end_year=2008)
-        self.assertIsNotNone(player_season_totals)
-
-    def test_2009_player_season_totals(self):
-        player_season_totals = client.players_season_totals(season_end_year=2009)
-        self.assertIsNotNone(player_season_totals)
-
-    def test_2010_player_season_totals(self):
-        player_season_totals = client.players_season_totals(season_end_year=2010)
-        self.assertIsNotNone(player_season_totals)
-
-    def test_2011_player_season_totals(self):
-        player_season_totals = client.players_season_totals(season_end_year=2011)
-        self.assertIsNotNone(player_season_totals)
-
-    def test_2012_player_season_totals(self):
-        player_season_totals = client.players_season_totals(season_end_year=2012)
-        self.assertIsNotNone(player_season_totals)
-
-    def test_2013_player_season_totals(self):
-        player_season_totals = client.players_season_totals(season_end_year=2013)
-        self.assertIsNotNone(player_season_totals)
-
-    def test_2014_player_season_totals(self):
-        player_season_totals = client.players_season_totals(season_end_year=2014)
-        self.assertIsNotNone(player_season_totals)
-
-    def test_2015_player_season_totals(self):
-        player_season_totals = client.players_season_totals(season_end_year=2015)
-        self.assertIsNotNone(player_season_totals)
-
-    def test_2016_player_season_totals(self):
-        player_season_totals = client.players_season_totals(season_end_year=2016)
-        self.assertIsNotNone(player_season_totals)
-
-    def test_2017_player_season_totals(self):
-        player_season_totals = client.players_season_totals(season_end_year=2017)
-        self.assertIsNotNone(player_season_totals)
-
-    def test_2018_player_season_totals(self):
-        player_season_totals = client.players_season_totals(season_end_year=2018)
-        self.assertIsNotNone(player_season_totals)
-
-    def test_2019_player_season_totals(self):
-        player_season_totals = client.players_season_totals(season_end_year=2019)
-        self.assertIsNotNone(player_season_totals)
-
     def test_2019_player_advanced_season_totals(self):
         player_season_totals = client.players_advanced_season_totals(season_end_year=2019)
         self.assertIsNotNone(player_season_totals)
